+++ src/main/java/me/inf32768/ultimate_scaler/option/UltimateScalerOptions.java
@@ -3,11 +3,7 @@ package me.inf32768.ultimate_scaler.option;
 import me.inf32768.ultimate_scaler.UltimateScaler;
 import me.inf32768.ultimate_scaler.shadowed.com.moandjiezana.toml.Toml;
 import net.fabricmc.loader.api.FabricLoader;
-import net.minecraft.text.OrderedText;
-import net.minecraft.text.Style;
 import net.minecraft.text.Text;
-import net.minecraft.text.TextContent;
-import net.minecraft.util.Language;
 import org.lwjgl.glfw.GLFW;
 
 import java.io.IOException;
@@ -15,7 +11,6 @@ import java.math.BigDecimal;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Arrays;
-import java.util.List;
 
 public final class UltimateScalerOptions {
     //Don't let anyone instantiate this class
@@ -23,7 +18,7 @@ public final class UltimateScalerOptions {
 
     public static ConfigImpl config;
     public static final Path CONFIG_PATH = FabricLoader.getInstance().getConfigDir().resolve("ultimate_scaler.toml");
-    public static final int CONFIG_VERSION = 1;
+    public static final int CONFIG_VERSION = 2;
 
     public static class ConfigImpl {
         public BigDecimal[] globalBigDecimalOffset = {BigDecimal.ZERO, BigDecimal.ZERO, BigDecimal.ZERO};
@@ -37,9 +32,14 @@ public final class UltimateScalerOptions {
         public int maxNoiseLogarithmValue = 7;
         public boolean extraYOffset = false;
         public boolean bigIntegerRewrite = false;
-        public boolean simulateEndRings = false;
+        public boolean fixEndRings = false;
         public boolean fixChunkGenerationOutOfBound = true;
         public boolean expandDatapackValueRange = true;
+        public boolean replaceDefaultFluid = false;
+        public String replaceDefaultFluidBlock = "minecraft:air";
+        public boolean replaceUndergroundLava = false;
+        public String replaceUndergroundLavaBlock = "minecraft:air";
+        public boolean publicTerrainPos = true;
     }
 
     static {
@@ -89,11 +89,18 @@ public final class UltimateScalerOptions {
         ConfigManager.writeEntry(CONFIG_PATH, "maintainPrecisionCustomDivisor", config.maintainPrecisionCustomDivisor, new String[] {Text.translatable("ultimate_scaler.options.worldgen.maintainPrecisionCustomDivisor").getString(), Text.translatable("ultimate_scaler.options.worldgen.maintainPrecisionCustomDivisor.tooltip").getString()});
         ConfigManager.writeEntry(CONFIG_PATH, "limitReturnValue", config.limitReturnValue, new String[] {Text.translatable("ultimate_scaler.options.worldgen.limitReturnValue").getString()});
         ConfigManager.writeEntry(CONFIG_PATH, "maxNoiseLogarithmValue", config.maxNoiseLogarithmValue, new String[] {Text.translatable("ultimate_scaler.options.worldgen.maxNoiseLogarithmValue").getString(), Text.translatable("ultimate_scaler.options.worldgen.maxNoiseLogarithmValue.tooltip").getString()});
+        ConfigManager.writeEntry(CONFIG_PATH, "replaceDefaultFluid", config.replaceDefaultFluid, new String[] {Text.translatable("ultimate_scaler.options.worldgen.replaceDefaultFluid").getString(), Text.translatable("ultimate_scaler.options.worldgen.replaceDefaultFluid.tooltip").getString()});
+        ConfigManager.writeEntry(CONFIG_PATH, "replaceDefaultFluidBlock", config.replaceDefaultFluidBlock, new String[] {});
+        ConfigManager.writeEntry(CONFIG_PATH, "replaceUndergroundLava", config.replaceUndergroundLava, new String[] {Text.translatable("ultimate_scaler.options.worldgen.replaceUndergroundLava").getString(), Text.translatable("ultimate_scaler.options.worldgen.replaceUndergroundLava.tooltip").getString()});
+        ConfigManager.writeEntry(CONFIG_PATH, "replaceUndergroundLavaBlock", config.replaceUndergroundLavaBlock, new String[] {});
         ConfigManager.writeEntry(CONFIG_PATH, "extraYOffset", config.extraYOffset, new String[] {Text.translatable("ultimate_scaler.options.worldgen.extraYOffset").getString(), Text.translatable("ultimate_scaler.options.worldgen.extraYOffset.tooltip").getString()});
         ConfigManager.writeEntry(CONFIG_PATH, "bigIntegerRewrite", config.bigIntegerRewrite, new String[] {Text.translatable("ultimate_scaler.options.worldgen.bigIntegerRewrite").getString(), Text.translatable("ultimate_scaler.options.worldgen.bigIntegerRewrite.tooltip.1").getString() + Text.translatable("ultimate_scaler.options.worldgen.bigIntegerRewrite.tooltip.2").getString() + Text.translatable("ultimate_scaler.options.worldgen.bigIntegerRewrite.tooltip.3").getString() + Text.translatable("ultimate_scaler.options.worldgen.bigIntegerRewrite.tooltip.4").getString()});
-        ConfigManager.writeEntry(CONFIG_PATH, "simulateEndRings", config.simulateEndRings, new String[] {Text.translatable("ultimate_scaler.options.worldgen.simulateEndRings").getString(), Text.translatable("ultimate_scaler.options.worldgen.simulateEndRings.tooltip").getString()});
-        ConfigManager.writeEntry(CONFIG_PATH, "fixChunkGenerationOutOfBound", config.fixChunkGenerationOutOfBound, new String[] {Text.translatable("ultimate_scaler.options.fix.chunkGenerationOutOfBound").getString(), Text.translatable("ultimate_scaler.options.fix.chunkGenerationOutOfBound.tooltip").getString()});
+        ConfigManager.writeEntry(CONFIG_PATH, "fixEndRings", config.fixEndRings, new String[] {Text.translatable("ultimate_scaler.options.worldgen.fixEndRings").getString(), Text.translatable("ultimate_scaler.options.worldgen.fixEndRings.tooltip").getString()});
+        if (UltimateScaler.MC_VERSION >= 4080) {
+            ConfigManager.writeEntry(CONFIG_PATH, "fixChunkGenerationOutOfBound", config.fixChunkGenerationOutOfBound, new String[] {Text.translatable("ultimate_scaler.options.fix.chunkGenerationOutOfBound").getString(), Text.translatable("ultimate_scaler.options.fix.chunkGenerationOutOfBound.tooltip").getString()});
+        }
         ConfigManager.writeEntry(CONFIG_PATH, "expandDatapackValueRange", config.expandDatapackValueRange, new String[] {Text.translatable("ultimate_scaler.options.fix.expandDatapackValueRange").getString(), Text.translatable("ultimate_scaler.options.fix.expandDatapackValueRange.tooltip.1").getString() + Text.translatable("ultimate_scaler.options.fix.expandDatapackValueRange.tooltip.2").getString()});
+        ConfigManager.writeEntry(CONFIG_PATH, "publicTerrainPos", config.publicTerrainPos, new String[] {Text.translatable("ultimate_scaler.options.server.publicTerrainPos").getString()});
     }
 
     public enum FarLandsPos {